Moving the 'cloudfoundry' package inside 'cloudproviders'
