+++ cmd/cluster-agent/api/v1/cloudfoundry_metadata.go
@@ -12,7 +12,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/DataDog/datadog-agent/pkg/util/cloudfoundry"
+	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/cloudfoundry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 	"github.com/gorilla/mux"
 )
