+++ pkg/tagger/collectors/garden_extract.go
@@ -3,7 +3,7 @@ package collectors
 import (
 	"fmt"
 
-	"github.com/DataDog/datadog-agent/pkg/util/cloudfoundry"
+	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/cloudfoundry"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
