+++ pkg/util/cloudproviders/cloudproviders.go
@@ -15,12 +15,12 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/hostname/kubelet"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
-	"github.com/DataDog/datadog-agent/pkg/util/cloudfoundry"
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/ecs"
 
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/alibaba"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/azure"
+	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/cloudfoundry"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/gce"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/tencent"
 )
