+++ pkg/util/containers/collectors/cloudfoundry.go
@@ -6,7 +6,7 @@
 package collectors
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/util/cloudfoundry"
+	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/cloudfoundry"
 	"github.com/DataDog/datadog-agent/pkg/util/containers"
 )
